right tridioagonal matrix made

+++ Koshi_giu/Cauchy_solver/Eiler.cs
@@ -11,12 +11,12 @@ namespace Cauchy
     class Eiler_method:Cauchy_Solver
     {
 
-        public override double [ ] Solve( Func<double, double [ ], double> [ ] right_funcs, Value [ ] prev_value, double x )
+        public override decimal [ ] Solve( Func<decimal, decimal [ ], decimal> [ ] right_funcs, Value [ ] prev_value, decimal x )
         {
             int dim = prev_value [ 3 ].U.Length;
-            double h = x - prev_value [ 3 ].X;
+            decimal h = x - prev_value [ 3 ].X;
 
-            double [ ] res = new double [ dim ];
+            decimal [ ] res = new decimal [ dim ];
             for ( int p = 0; p < dim; ++p )
                 res [ p ] = prev_value[0].U[ p ] + h * right_funcs [ p ](prev_value[0].X, prev_value[0].U );
             return res;
