+++ Koshi_giu/Tasks/Cauchy_Solver.cs
@@ -11,7 +11,7 @@ namespace Cauchy
     abstract class Cauchy_Solver
     {
         protected int dim ;
-        protected double h;
-        public abstract double [ ] Solve ( Func<double, double [ ], double> [ ] right_funcs, Value[] prev_value ,double x );
+        protected decimal h;
+        public abstract decimal [ ] Solve ( Func<decimal, decimal [ ], decimal> [ ] right_funcs, Value[] prev_value ,decimal x );
     }
 }
